refactor(direct-control): extract Tuner health wrapper
Refs: openspec/changes/civ7-support-direct-control-modularization

Move only the Tuner health wrapper orchestration, generated Tuner health command, command-result parser, and internal session helper into packages/civ7-direct-control/src/runtime/tuner-health.ts while keeping public facade call-through in packages/civ7-direct-control/src/index.ts.

The facade still owns session creation/close, reconnect execution, and readiness wait orchestration. The moved helper receives session-command execution as an injected dependency so the runtime module does not import executable facade values.

Tuner health remains internal service/runtime-status support: normal CLI projection is omitted/debug-only or left for later semantic state-machine surfaces, and debug service projection is raw Tuner health/readiness diagnostics. Playable status, bounded root inspection, capability catalog, telemetry, hotseat runtime proof, AI ingestion, CLI semantic projection, and Effect/oRPC procedure-core work remain pending.

Verification:
- bun run --cwd packages/civ7-direct-control test -- runtime-and-catalog.test.ts restart-lifecycle.test.ts setup-and-lifecycle.test.ts
- bun run --cwd packages/civ7-direct-control check
- bun run --cwd packages/civ7-direct-control test
- bun run --cwd packages/civ7-direct-control build
- bun run openspec -- validate civ7-support-direct-control-modularization --strict
- bun run resources:status
- bun run check:cli
- bun run test:cli:play
- git diff --check
- ownership and no-cycle scans for Tuner health helpers

Local package/source relocation proof only; no runtime or live-game proof claimed.
